// src/app/database/entities/rrhh/notificacion-rrhh-enums.ts
export enum TipoNotificacionRrhh {
  PRESTAMO_VENCIDO = 'PRESTAMO_VENCIDO',
  CUOTA_VENCIDA = 'CUOTA_VENCIDA',
  CUMPLEANIOS = 'CUMPLEANIOS',
  VACACION_PROXIMA = 'VACACION_PROXIMA',
  CONTRATO_VENCE = 'CONTRATO_VENCE',
  LIQUIDACION_PENDIENTE = 'LIQUIDACION_PENDIENTE',
  COMISION_PENDIENTE = 'COMISION_PENDIENTE',
  DOCUMENTO_VENCE = 'DOCUMENTO_VENCE',
}

export enum PrioridadNotificacion {
  ALTA = 'ALTA',
  MEDIA = 'MEDIA',
  BAJA = 'BAJA',
}